Move common code to helper
